fix: Change score type to Int64 in GameStatsView
Updates the `score` property in `GameStatsView` from `Int32` to `Int64` to handle larger score values.

+++ iosApp/iosApp/Views/game/GameView.swift
@@ -153,7 +153,7 @@ private typealias ErrorChild = GameComponentDialogChildError
 
 
 struct GameStatsView: View {
-    let score: Int32
+    let score: Int64
     let lines: Int32
     let time: Int64
